scripts/pyproject.toml
[project]
name = "homelab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
